Refactor bot utils
