+++ AGENTS.md
@@ -1,11 +1,11 @@
 # AGENTS Instructions
 
 This repository is a simple Telegram bot powered by Pyrogram and OpenAI.
+The main logic now lives in `app.py` and helper functions are in `bot_utils.py`.
 
 * Whenever you modify any Python file, especially `app.py`, run
-  `python -m py_compile app.py` to ensure there are no syntax errors.
+  `python -m py_compile app.py bot_utils.py` to ensure there are no syntax errors.
 * Use four spaces for indentation.
 * Keep commit messages brief but descriptive.
 * In pull request summaries, mention the key code changes and summarize
   the result of running the above command.
-
